Add BaseTest for test boilerplate
