+++ shared/src/androidTest/kotlin/co/touchlab/kmp/BaseTest.kt
@@ -1,0 +1,17 @@
+package co.touchlab.kampstarter
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.runBlocking
+import org.junit.Rule
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+actual abstract class BaseTest {
+    @get:Rule
+    var coroutineTestRule = CoroutineTestRule()
+
+    actual fun <T> runTest(block: suspend CoroutineScope.() -> T) {
+        runBlocking { block() }
+    }
+}
