+++ shared/src/commonTest/kotlin/co/touchlab/kampstarter/BaseTest.kt
@@ -1,0 +1,7 @@
+package co.touchlab.kampstarter
+
+import kotlinx.coroutines.CoroutineScope
+
+expect abstract class BaseTest(){
+    fun <T> runTest(block: suspend CoroutineScope.() -> T)
+}
