+++ shared/src/iosTest/kotlin/co/touchlab/kmp/BaseTest.kt
@@ -1,0 +1,26 @@
+package co.touchlab.kampstarter
+
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.launch
+import platform.CoreFoundation.CFRunLoopGetCurrent
+import platform.CoreFoundation.CFRunLoopRun
+import platform.CoreFoundation.CFRunLoopStop
+
+actual abstract class BaseTest {
+    actual fun <T> runTest(block: suspend CoroutineScope.() -> T) {
+        var error: Throwable? = null
+        GlobalScope.launch(Dispatchers.Main) {
+            try {
+                block()
+            } catch (t: Throwable) {
+                error = t
+            } finally {
+                CFRunLoopStop(CFRunLoopGetCurrent())
+            }
+        }
+        CFRunLoopRun()
+        error?.also { throw it }
+    }
+}
